+++ NEWS.md
@@ -1,3 +1,21 @@
+# CFtime 1.5.0
+
+* Code is updated to align with the CF 1.12 release. Specifically, calendars 
+`utc` and `tai` are added.
+* `standard` calendar now uses mixed Gregorian/Julian calendar as defined in the
+CF Metadata Conventions. `proleptic_gregorian` is now a separate calendar with
+its own code base.
+* Negative offsets from a calendar origin are allowed.
+* Code is refactored to R6. R6 class CFTime replaces S4 class CFtime (note the
+difference in case). S4 class CFdatum has been replaced by hierarchy of
+R6 CFCalendar classes, with various non-exported functions converted into
+methods of CFCalendar. The code is now much cleaner and easier to extend, such
+as with the two new calendars `utc` and `tai`.
+* ISO8601 timestamps can use a comma "," as decimal mark to separate a 
+fractional part from a time element, in addition to the dot ".".
+* Do not drop degenerate dimension on bounds when only 1 offset is included in
+subsetting.
+
 # CFtime 1.4.1
 
 * Method `slab()` has an additional argument `rightmost.closed` to indicate if